Refactor describable.py & explainable.py
